Add `docker` to the container
This way, we can build images inside of the Docker container itself. I
also wanted it to be able to run images too, but that was a bit harder.
To do it, you either have to do a Docker in Docker, which is not
recommended, or mount the `docker.sock` file into the container at the
correct location (which is `/var/run/docker`). But the latter had the
issue that is was mounted with the wrong group ownership. It is mounted
with the group ownership of the host, which in my case is GID 139. The
docker group in the container, however, is GID 999. This conflict and
therefor, the default developer user couldn't run Docker commands. I
could, ofcourse, change the GID inside the coutainer to 139, but that
would remove all portability.

Instead, I choose to add a alias for `docker` in the `.bashrc` file.
This alias simply runs `sudo docker`, which makes it work again. It
might not be ideal but it works.

Fixes #26